+++ appFrontend/src/navigation/AuthNavigation.tsx
@@ -3,6 +3,7 @@ import { AuthStackStackParamList } from './types/navigation';
 import LandingScreen from '@/screens/landingScreen';
 import SignInScreen from '@/screens/signInScreen';
 import SignUpScreen from '@/screens/signUpScreen';
+import SocialLoginScreen from '@/screens/landingScreen/components/SocialLoginScreen';
 import { useNavigation } from '@react-navigation/native';
 
 const Stack = createStackNavigator<AuthStackStackParamList>();
@@ -17,6 +18,7 @@ const AuthNavigation = () => {
       <Stack.Screen name="Landing" component={LandingScreen} />
       <Stack.Screen name="SignIn" component={SignInScreen} />
       <Stack.Screen name="SignUp" component={SignUpScreen} />
+      <Stack.Screen name="SocialLogin" component={SocialLoginScreen} />
     </Stack.Navigator>
   );
 };
