+++ appFrontend/src/navigation/types/navigation.ts
@@ -1,4 +1,5 @@
 import { Path, SubPath } from '@/global/apis/entity';
+import { SocialLoginType } from '@/screens/landingScreen/components/SocialLoginButtons';
 
 export type RootStackParamList = {
   AuthStack: { screen: 'Landing' };
@@ -15,6 +16,7 @@ export type AuthStackStackParamList = {
   Landing: undefined;
   SignIn: undefined;
   SignUp: undefined;
+  SocialLogin: { socialLoginType: SocialLoginType };
 };
 
 export type HomeStackParamList = {
